+++ src/main/java/eu/dissco/core/handlemanager/domain/requests/MediaObjectRequest.java
@@ -1,6 +1,7 @@
 package eu.dissco.core.handlemanager.domain.requests;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import eu.dissco.core.handlemanager.domain.requests.helpers.PhysicalIdentifier;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
