+++ src/test/java/eu/dissco/core/handlemanager/testUtils/TestUtils.java
@@ -36,13 +36,13 @@ import eu.dissco.core.handlemanager.domain.jsonapi.JsonApiLinks;
 import eu.dissco.core.handlemanager.domain.jsonapi.JsonApiWrapperRead;
 import eu.dissco.core.handlemanager.domain.jsonapi.JsonApiWrapperWrite;
 import eu.dissco.core.handlemanager.domain.repsitoryobjects.HandleAttribute;
-import eu.dissco.core.handlemanager.domain.requests.DigitalOrPhysical;
+import eu.dissco.core.handlemanager.domain.requests.helpers.DigitalOrPhysical;
 import eu.dissco.core.handlemanager.domain.requests.DigitalSpecimenBotanyRequest;
 import eu.dissco.core.handlemanager.domain.requests.DigitalSpecimenRequest;
 import eu.dissco.core.handlemanager.domain.requests.DoiRecordRequest;
 import eu.dissco.core.handlemanager.domain.requests.HandleRecordRequest;
-import eu.dissco.core.handlemanager.domain.requests.PhysicalIdentifier;
-import eu.dissco.core.handlemanager.domain.requests.PreservedOrLiving;
+import eu.dissco.core.handlemanager.domain.requests.helpers.PhysicalIdentifier;
+import eu.dissco.core.handlemanager.domain.requests.helpers.PreservedOrLiving;
 import eu.dissco.core.handlemanager.domain.requests.TombstoneRecordRequest;
 import java.io.StringWriter;
 import java.nio.charset.StandardCharsets;
